feat(bsw): port the long-read 8-bit recovery to AVX2 and AVX-512
Apply the diagonal-offset int8 position encoding + per-lane score re-baselining
(landed for NEON/SSE smithWaterman128_8) to smithWaterman256_8 (AVX2, 32 lanes)
and smithWaterman512_8 (AVX-512BW, 64 lanes), so reads >=128bp keep the widest
8-bit path on x86. AVX-512 uses __mmask64 + _mm512_mask_blend_epi8; AVX2 uses
_mm256_blendv_epi8.

+++ src/bandedSWA.h
@@ -355,8 +355,8 @@ public:
 
     void smithWaterman256_8(uint8_t seq1SoA[],
                             uint8_t seq2SoA[],
-                            uint8_t nrow,
-                            uint8_t ncol,
+                            int nrow,
+                            int ncol,
                             SeqPair *p,
                             uint8_t h0[],
                             uint16_t tid,
@@ -415,8 +415,8 @@ public:
 
     void smithWaterman512_8(uint8_t seq1SoA[],
                             uint8_t seq2SoA[],
-                            uint8_t nrow,
-                            uint8_t ncol,
+                            int nrow,
+                            int ncol,
                             SeqPair *p,
                             uint8_t h0[],
                             uint16_t tid,
